typos, add SublimeText link
Sublime Text is AFAIK not IDE, only editor so I updated that sentence.

Sublime Text is also commercial - I'd like to propose to replace ST with Atom, which is very similar in extensibility, but released as Open Source.

Also, in comments you rejected to add NetBeans, because of its bad performance and UI, but both Eclipse and phpDesigner are from my experience much worse. Brackets is more for front-end development.